requirements.txt
django==1.7.1
MySQL-python==1.2.5
django-flat-theme
python-social-auth==0.2.7
easy_thumbnails==2.0.1
django-image-cropping==0.8.2
django-tagging==0.4
django-annoying==0.8.1
django-contrib-comments==1.5
django-decouple==2.1
django-easy-pdf==0.1.0
git+https://github.com/chrisglass/xhtml2pdf.git
django-widget-tweaks==1.3
django-ckeditor==4.4.4
django-taggit==0.15.0
django-daterange-filter==1.1.1